#909 refactor(agent): canonicalize Host route projection
