+++ packages/agent/src/server/__tests__/registerAgentRoutes.test.ts
@@ -1109,14 +1109,17 @@ test('createAgentApp has zero runtime imports from @hachej/boring-core', async (
   expect(matches).toBeNull()
 })
 
-test('registerAgentRoutes awaits the Agent Host funnel and contains no local construction path', async () => {
+test('registerAgentRoutes normalizes once and delegates the complete profile through the Agent Host', async () => {
   const source = await readFile(join(import.meta.dirname, '..', 'registerAgentRoutes.ts'), 'utf8')
+  const policySource = await readFile(join(import.meta.dirname, '..', 'agentHostLegacyRoutePolicy.ts'), 'utf8')
   const hostSource = await readFile(join(import.meta.dirname, '..', 'agent-host', 'createAgentHost.ts'), 'utf8')
 
   expect(source.match(/\bcreateAgentHost\s*\(/g)).toHaveLength(1)
-  expect(source).toMatch(/agentHost\s*=\s*opts\.agentHost\?\.created\s*\?\?\s*await createAgentHost\s*\(/)
-  expect(source).toMatch(/await resolveAgentHostCompatibilityComposition\s*\(/)
-  expect(source).not.toMatch(/\b(?:buildAgentComposition|createAgentRuntimeBridge|createCompositionRuntimeBridge|buildHarnessAgentTools|buildFilesystemAgentTools|buildUploadAgentTools|createPiCodingAgentHarness)\s*\(/)
+  expect(source).toMatch(/created\s*=\s*opts\.agentHost\?\.created\s*\?\?\s*await createAgentHost\s*\(/)
+  expect(source).toMatch(/await app\.register\s*\(\s*created\.registerRoutes\s*\(/)
+  expect(source).not.toMatch(/await resolveAgentHostCompatibilityComposition\s*\(/)
+  expect(policySource).toMatch(/await resolveAgentHostCompatibilityComposition\s*\(/)
+  expect(`${source}\n${policySource}`).not.toMatch(/\b(?:createAgentRuntimeBridge|createCompositionRuntimeBridge|buildHarnessAgentTools|buildFilesystemAgentTools|buildUploadAgentTools|createPiCodingAgentHarness)\s*\(/)
   expect(hostSource.match(/await buildAgentComposition\s*\(/g)).toHaveLength(1)
 })
