+++ packages/agent/src/server/index.ts
@@ -138,10 +138,14 @@ export type {
   AgentEffectAdmission,
   AgentFleetCompiler,
   AgentGatewayEffect,
+  AgentHostAddressedHttpProjectionOptions,
   AgentHostAgentSpec,
   AgentHostDescription,
   AgentHostHandle,
   AgentHostHttpProjectionOptions,
+  AgentHostLegacyHttpProjectionOptions,
+  AgentHostLegacyRoutePolicy,
+  AgentHostLegacyRoutePolicyMountInput,
   AgentRequestKey,
   AgentRequestLedger,
   AgentRequestLedgerRecord,
@@ -172,7 +176,12 @@ export {
   PI_PACKAGE_RESOURCE_FILTERS,
 } from './piPackages'
 export { registerAgentRoutes } from './registerAgentRoutes'
+export {
+  createAgentHostLegacyRoutePolicy,
+} from './agentHostLegacyRoutePolicy'
 export type {
+  AgentHostLegacyRoutePolicyOptions,
+  AgentHostLegacyRouteScopePolicy,
   PrebuiltAgentHostRoutePolicy,
   RegisterAgentRoutesOptions,
 } from './registerAgentRoutes'
