+++ packages/core/src/app/server/__tests__/createCoreWorkspaceAgentServer.telemetry.test.ts
@@ -29,10 +29,13 @@ const dbMock = vi.hoisted(() => {
 vi.mock('@hachej/boring-agent/server', () => ({
   autoDetectMode: () => 'direct',
   compactPiPackages: (packages: unknown[]) => packages,
-  createAgentHost: async () => ({ marker: 'prebuilt-agent-host' }),
-  registerAgentRoutes: async (_app: unknown, opts: Record<string, unknown>) => {
-    agentMock.registerOptions.push(opts)
-  },
+  createAgentHostLegacyRoutePolicy: (options: Record<string, unknown>) => ({ options }),
+  createAgentHost: async () => ({
+    marker: 'prebuilt-agent-host',
+    registerRoutes: (projection: { legacyRoutePolicy: { options: Record<string, unknown> } }) => async () => {
+      agentMock.registerOptions.push(projection.legacyRoutePolicy.options)
+    },
+  }),
 }))
 
 vi.mock('@hachej/boring-workspace/app/server', () => ({
