+++ packages/core/src/app/server/createCoreWorkspaceAgentServer.ts
@@ -6,9 +6,9 @@ import {
   compactPiPackages,
   autoDetectMode,
   createAgentHost,
+  createAgentHostLegacyRoutePolicy,
   createRemoteWorkerModeAdapter,
   provisionWorkspaceRuntime,
-  registerAgentRoutes,
   type AgentEffectAdmission,
   type AgentFleetCompiler,
   type AgentHostAgentSpec,
@@ -1213,12 +1213,7 @@ export async function createCoreWorkspaceAgentServer(
     },
   })
 
-  await app.register(registerAgentRoutes, {
-    agentHost: {
-      created: agentHost,
-      defaultAgentTypeId,
-      issueScope: scopeAuthority.issueScope,
-    },
+  const legacyRoutePolicy = createAgentHostLegacyRoutePolicy({
     workspaceRoot,
     sessionId: options.sessionId,
     templatePath: options.templatePath,
@@ -1228,8 +1223,8 @@ export async function createCoreWorkspaceAgentServer(
     runtimeModeAdapter,
     runtimeHost,
     version: options.version,
-    // The prebuilt Host applies effectAdmission to AgentGateway effects;
-    // registerAgentRoutes keeps admitEffect on reload and command routes only.
+    // The Host applies effectAdmission to AgentGateway effects; the normalized
+    // legacy policy keeps admitEffect on reload and command routes only.
     admitEffect: options.admitEffect,
     extraTools: [
       ...(options.extraTools ?? []),
@@ -1299,7 +1294,14 @@ export async function createCoreWorkspaceAgentServer(
       ...(options.runtimeEnvContributions ?? []),
       ...(coreBridge.runtimeEnvContribution ? [coreBridge.runtimeEnvContribution] : []),
     ],
+  }, {
+    issueScope: scopeAuthority.issueScope,
   })
+
+  await app.register(agentHost.registerRoutes({
+    defaultAgentTypeId,
+    legacyRoutePolicy,
+  }))
 
   await app.register(uiRoutes, {
     getWorkspaceId: resolveWorkspaceId,
